requirements.txt
streamlit==1.31.0
snowflake-connector-python==3.7.0
pandas==2.2.0
numpy==1.26.3
plotly==5.18.0
python-dotenv==1.0.1
gspread==5.12.4
oauth2client==4.1.3
openpyxl==3.1.2
streamlit-authenticator==0.3.1
pyyaml==6.0.1
bcrypt==4.1.2
extra-streamlit-components==0.1.71
